requirements.txt
requests_oauthlib
requests
python-dotenv
pyopenssl